+++ changelog.md
@@ -1,6 +1,7 @@
 v0.0.11-alpha
 -------------
 * Update research table project background
+* Made Flint and Steel set Treefolk on fire
 
 v0.0.10-alpha
 -------------